requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
boto3==1.24.84
botocore==1.27.84
Brotli==1.0.9
cairocffi==1.4.0
CairoSVG==2.5.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
cssselect2==0.6.0
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.0
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.9.0
django-cors-headers==3.13.0
django-object-actions==4.0.0
django-password-validators==1.7.0
django-sendgrid-v5==1.2.1
django-storages==1.13.1
django-xhtml2pdf==0.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
fonttools==4.37.4
future==0.18.2
html5lib==1.1
idna==3.4
inflect==6.0.0
jmespath==1.0.1
lxml==4.9.1
numpy==1.24.1
oscrypto==1.3.0
pandas==1.5.2
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.10.2
pydyf==0.5.0
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
PyPDF3==1.0.6
pyphen==0.13.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
s3transfer==0.6.0
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
svglib==1.4.1
tinycss2==1.1.1
tqdm==4.64.1
typing-extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
WeasyPrint==52.5
webencodings==0.5.1
wget==3.2
whitenoise==6.2.0
wkhtmltopdf==0.2
xhtml2pdf==0.2.8
zopfli==0.2.1
